Refactor utils into own module
